refactor: 유저 관련 리팩토링
