+++ src/main/java/com/roome/roome/be/domain/product/controller/ProductController.java
@@ -2,8 +2,6 @@ package com.roome.roome.be.domain.product.controller;
 
 import java.util.List;
 
-import com.roome.roome.be.domain.product.dto.response.ProductToggleLikeResponse;
-import com.roome.roome.be.domain.product.dto.response.ProductToggleScrapResponse;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -132,43 +130,5 @@ public class ProductController {
 	) {
 		var page = productService.getList(shopId, category, colorTags, materialTags, styleTags, featureTags, moodTags, match, keyWord, minPrice, maxPrice, pageable);
 		return ApiResponse.success(SuccessStatus.GET_PRODUCT_LIST, page);
-	}
-
-	// 상품 좋아요 기능 구현
-	@PostMapping("{productId}/like")
-	@Operation(
-			summary = "상품 좋아요 토글",
-			description = "이미 좋아요가 눌려 있으면 취소하고, 눌려 있지 않으면 좋아요를 추가합니다."
-	)
-	@Parameters({
-			@Parameter(name = "productId", description = "좋아요를 누를 상품의 ID", example = "123"),
-	})
-	@io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "좋아요 토글 성공", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ProductToggleLikeResponse.class)))
-	@io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "404", description = "유저가 존재하지 않거나 상품이 존재하지 않음", content = @Content(mediaType = "application/json"))
-	public ResponseEntity<ApiResponse<ProductToggleLikeResponse>> toggleProductLike(
-			@PathVariable("productId") Long productId,
-			@AuthenticationPrincipal Long userId
-	) {
-		ProductToggleLikeResponse response = productService.toggleProductLike(productId, userId);
-		return ApiResponse.success(SuccessStatus.CREATE_PRODUCT_LIKE,response);
-	}
-
-	// 상품 스크랩 기능 구현
-	@PostMapping("/{productId}/scrap")
-	@Operation(
-			summary = "상품 스크랩 토글",
-			description = "이미 스크랩이 되어 있으면 취소하고, 되어 있지 않으면 스크랩에 추가합니다."
-	)
-	@Parameters({
-			@Parameter(name = "productId", description = "스크랩할 상품의 ID", example = "123"),
-	})
-	@io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "스크랩 토글 성공", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ProductToggleScrapResponse.class)))
-	@io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "404", description = "유저가 존재하지 않거나 상품이 존재하지 않음", content = @Content(mediaType = "application/json"))
-	public ResponseEntity<ApiResponse<ProductToggleScrapResponse>> toggleProductScrap(
-			@PathVariable("productId") Long productId,
-			@AuthenticationPrincipal Long userId
-	) {
-		ProductToggleScrapResponse response = productService.toggleProductScrap(productId, userId);
-		return ApiResponse.success(SuccessStatus.CREATE_PRODUCT_LIKE,response);
 	}
 }
