+++ src/main/java/com/roome/roome/be/domain/reference/service/ReferenceService.java
@@ -7,13 +7,11 @@ import com.roome.roome.be.common.s3.service.S3Service;
 import com.roome.roome.be.common.status.ErrorStatus;
 import com.roome.roome.be.domain.reference.dto.response.CommonReferenceInfo;
 import com.roome.roome.be.domain.reference.dto.response.ReferenceListResponse;
-import com.roome.roome.be.domain.reference.dto.response.ReferenceToggleScrapResponse;
 import com.roome.roome.be.domain.reference.entity.Reference;
 import com.roome.roome.be.domain.reference.entity.ReferenceImage;
 import com.roome.roome.be.domain.reference.repository.ReferenceImageRepository;
 import com.roome.roome.be.domain.reference.repository.ReferenceRepository;
 import com.roome.roome.be.domain.user.entity.User;
-import com.roome.roome.be.domain.user.entity.UserScrapReference;
 import com.roome.roome.be.domain.user.repository.UserRepository;
 import com.roome.roome.be.domain.user.repository.UserScrapReferenceRepository;
 import lombok.RequiredArgsConstructor;
@@ -22,7 +20,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -60,30 +57,6 @@ public class ReferenceService {
                 })
                 .toList();
         referenceImageRepository.saveAll(referenceImageList);
-    }
-
-    public ReferenceToggleScrapResponse toggleReferenceScrap(Long referenceId, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.USER_NOT_FOUND));
-        Reference reference = findReferenceById(referenceId);
-
-        boolean scrapped;
-
-        Optional<UserScrapReference> existing = userScrapReferenceRepository.findByUserAndReference(user,reference);
-        if (existing.isEmpty()) {
-            UserScrapReference userScrapReference = UserScrapReference.builder()
-                    .user(user)
-                    .reference(reference)
-                    .build();
-            userScrapReferenceRepository.save(userScrapReference);
-            scrapped = true;
-        }
-        else {
-            userScrapReferenceRepository.delete(existing.get());
-            scrapped = false;
-        }
-
-        return new ReferenceToggleScrapResponse(scrapped);
     }
 
     public ReferenceListResponse getReferenceList(Long userId) {
