+++ src/main/java/com/roome/roome/be/domain/user/service/UserService.java
@@ -8,15 +8,9 @@ import com.roome.roome.be.common.status.ErrorStatus;
 import com.roome.roome.be.domain.auth.dto.response.CheckEmailResponse;
 import com.roome.roome.be.domain.auth.dto.response.CheckNicknameResponse;
 import com.roome.roome.be.domain.auth.dto.request.SignUpRequest;
-import com.roome.roome.be.domain.product.dto.response.CommonProductInfo;
-import com.roome.roome.be.domain.product.entity.Product;
-import com.roome.roome.be.domain.reference.dto.response.CommonReferenceInfo;
 import com.roome.roome.be.domain.user.dto.request.UpdateUserProfileRequest;
-import com.roome.roome.be.domain.user.dto.request.UserOnboardingRequest;
 import com.roome.roome.be.domain.user.dto.response.*;
 import com.roome.roome.be.domain.user.entity.User;
-import com.roome.roome.be.domain.user.entity.UserOnboarding;
-import com.roome.roome.be.domain.user.entity.UserView;
 import com.roome.roome.be.domain.user.enums.LoginType;
 import com.roome.roome.be.domain.user.enums.Role;
 import com.roome.roome.be.domain.user.repository.*;
@@ -27,7 +21,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -35,19 +28,12 @@ import java.util.Optional;
 @Slf4j
 public class UserService {
     private final UserRepository userRepository;
-    private final UserOnboardingRepository userOnboardingRepository;
-    private final UserViewRepository userViewRepository;
-
-    private final UserLikeCustomRepositoryImpl userLikeCustomRepositoryImpl;
-    private final UserScrapProductCustomRepositoryImpl userScrapProductCustomRepositoryImpl;
-    private final UserViewCustomRepositoryImpl userViewCustomRepositoryImpl;
-    private final UserScrapReferenceCustomRepositoryImpl scrapReferenceCustomRepositoryImpl;
 
     private final S3Service s3Service;
     private final ImageUrlBuilder imageUrlBuilder;
 
     public UserProfileResponse getUserProfile(Long userId) {
-        User user = findUserById(userId);
+        User user = getUserById(userId);
         return new UserProfileResponse(
                 user.getId(),
                 user.getProfileImage(),
@@ -59,7 +45,7 @@ public class UserService {
 
     @Transactional
     public void updateUserProfile(Long userId, UpdateUserProfileRequest request){
-        User user = findUserById(userId);
+        User user = getUserById(userId);
 
         if(request.nickname() != null && !request.nickname().isBlank())
             user.updateNickname(request.nickname());
@@ -69,91 +55,6 @@ public class UserService {
             String objectKey = s3Service.uploadObject(StorageScope.USER_PROFILE,userId,file);
             user.updateProfileImage(imageUrlBuilder.build(objectKey));
         }
-    }
-
-    // 유저 온보딩 저장
-    @Transactional
-    public void saveUserOnboarding(Long userId, UserOnboardingRequest userOnboardingRequest) {
-        User user = findUserById(userId);
-        userOnboardingRepository.findByUser(user)
-                .ifPresentOrElse(
-                        UserOnboarding -> updateUserOnboarding(UserOnboarding, userOnboardingRequest),
-                        () -> registerUserOnboarding(user, userOnboardingRequest)
-                );
-    }
-
-    // 유저 온보딩 업데이트
-    private void updateUserOnboarding(UserOnboarding userOnboarding, UserOnboardingRequest userOnboardingRequest) {
-        userOnboarding.update(
-                userOnboardingRequest.ageGroup(),
-                userOnboardingRequest.gender(),
-                userOnboardingRequest.moodType(),
-                userOnboardingRequest.spaceType()
-        );
-    }
-
-    // 유저 온보딩 저장
-    private void registerUserOnboarding(User user, UserOnboardingRequest userOnboardingRequest) {
-        UserOnboarding newOnboarding = UserOnboarding.builder()
-                .user(user)
-                .ageGroup(userOnboardingRequest.ageGroup())
-                .gender(userOnboardingRequest.gender())
-                .moodType(userOnboardingRequest.moodType())
-                .spaceType(userOnboardingRequest.spaceType())
-                .build();
-
-        userOnboardingRepository.save(newOnboarding);
-    }
-
-    //유저온보딩 존재 여부
-    public UserOnboardingExistResponse checkExistence(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.USER_NOT_FOUND));
-
-        boolean exists = userOnboardingRepository.findByUser(user).isPresent();
-
-        return new UserOnboardingExistResponse(exists);
-    }
-
-    // 유저가 좋아요를 누른 상품 리스트 조회
-    public UserLikeProductListResponse getUserLikedProductList(Long userId) {
-        List<CommonProductInfo> userLikeProductList = userLikeCustomRepositoryImpl.findUserLikeProductListByUserId(userId);
-        return new UserLikeProductListResponse(userLikeProductList);
-    }
-
-    // 유저가 스크랩한 상품 리스트 조회
-    public UserScrappedProductListResponse getUserScrappedProductList(Long userId) {
-        List<CommonProductInfo> userScrappedProductList = userScrapProductCustomRepositoryImpl.findUserScrappedProductListByUserId(userId);
-        return new UserScrappedProductListResponse(userScrappedProductList);
-    }
-
-    // 유저가 스크랩한 레퍼런스 리스트 조회
-    public UserScrappedReferenceListResponse getUserScrappedReferenceList(Long userId) {
-
-        List<CommonReferenceInfo> rawList = scrapReferenceCustomRepositoryImpl
-                .findUserScrappedReferenceListByUserId(userId);
-
-        log.error("중간 점검 {}", rawList);
-        List<CommonReferenceInfo> finalList = rawList.stream()
-                .map(raw -> new CommonReferenceInfo(
-                        raw.referenceId(),
-                        raw.nickname(),
-                        raw.userId(),
-                        raw.imageUrlList().stream()
-                                .map(imageUrlBuilder::build)  // ★ 여기서 URL 변환
-                                .toList(),
-                        raw.scrapCount()
-                ))
-                .toList();
-
-        return new UserScrappedReferenceListResponse(finalList);
-    }
-
-
-    // 유저가 최근 본 상품 리스트 조회
-    public UserRecentViewedProductListResponse getUserRecentViewedProductList(Long userId) {
-        List<CommonProductInfo> userRecentViewedProductList = userViewCustomRepositoryImpl.findUserRecentViewedProductListByUserId(userId);
-        return new UserRecentViewedProductListResponse(userRecentViewedProductList);
     }
 
     // 이메일 중복 검사
@@ -189,37 +90,34 @@ public class UserService {
         userRepository.save(user);
     }
 
-    // User View 등록
-    public void registerUserView(Product product, User user) {
-        UserView userView = userViewRepository.findByUserAndProduct(user, product);
-
-        if (userView == null) {
-            userView = UserView.builder()
-                    .user(user)
-                    .product(product)
-                    .build();
-            userViewRepository.save(userView);
-        } else {
-            userView.touch();
-        }
-
-    }
-
     // 이메일로 유저 찾기
-    public User findUserByEmail(String email) {
+    public User getUserByEmail(String email) {
         return userRepository.findByEmail(email)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.EMAIL_NOT_FOUND_2));
     }
 
     // 아이디로 유저 찾기
-    public User findUserById(Long id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.USER_NOT_FOUND));
     }
 
     // 전화번호로 유저 찾기
-    public Optional<User> findUserByPhoneNumber(String phoneNumber) {
+    public Optional<User> getUserByPhoneNumber(String phoneNumber) {
         return userRepository.findUserByPhoneNumber(phoneNumber);
+    }
+
+    // RefreshToken으로 User 검색
+    public User getUserByRefreshToken(Claims claims, String refreshToken) {
+        String id = claims.getSubject();
+
+        User user = userRepository.findById(Long.parseLong(id))
+                .orElseThrow(() -> new GeneralException(ErrorStatus.REFRESH_TOKEN_NOT_FOUND));
+
+        if (!refreshToken.equals(user.getRefreshToken())) {
+            throw new GeneralException(ErrorStatus.REFRESH_TOKEN_MISMATCH);
+        }
+        return user;
     }
 
     // RefreshToken 업데이트
@@ -240,19 +138,6 @@ public class UserService {
     // 비밀번호 수정
     public void updatePassword(User user, String encryptedPassword) {
         user.updatePassword(encryptedPassword);
-    }
-
-    // RefreshToken으로 User 검색
-    public User findUserByRefreshToken(Claims claims, String refreshToken) {
-        String id = claims.getSubject();
-
-        User user = userRepository.findById(Long.parseLong(id))
-                .orElseThrow(() -> new GeneralException(ErrorStatus.REFRESH_TOKEN_NOT_FOUND));
-
-        if (!refreshToken.equals(user.getRefreshToken())) {
-            throw new GeneralException(ErrorStatus.REFRESH_TOKEN_MISMATCH);
-        }
-        return user;
     }
 
 }
